test(EuiDataGridCell): move EuiDataGrid column sorting jest tests to a separate file
