+++ packages/eui/src/components/datagrid/data_grid.test.tsx
@@ -110,52 +110,6 @@ function closeColumnSorterSelection(datagrid: ReactWrapper) {
   return columnSelectionPopover;
 }
 
-function getColumnSortDirection(
-  datagrid: ReactWrapper,
-  columnId: string
-): [ReactWrapper, string] {
-  // get the button that sorts by this column
-  let columnSorter = datagrid.find(
-    `div[data-test-subj="euiDataGridColumnSorting-sortColumn-${columnId}"]`
-  );
-  if (columnSorter.length === 0) {
-    // need to enable this column
-    openColumnSorterSelection(datagrid);
-
-    // find button to enable this column and click it
-    const selectColumnButton = datagrid.find(
-      `button[data-test-subj="dataGridColumnSortingPopoverColumnSelection-${columnId}"]`
-    );
-    expect(selectColumnButton.length).toBe(1);
-    // @ts-ignore onClick is known to exist, and does not require an argument in this usage
-    act(() => selectColumnButton.props().onClick());
-
-    // close column selection popover
-    closeColumnSorterSelection(datagrid);
-
-    // find the column sorter
-    const columnSelectionPopover = datagrid.find(
-      'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
-    );
-    columnSorter = columnSelectionPopover.find(
-      `div[data-test-subj="euiDataGridColumnSorting-sortColumn-${columnId}"]`
-    );
-  }
-
-  expect(columnSorter.length).toBe(1);
-  const activeSort = columnSorter.find(
-    'button[className*="euiButtonGroupButton-isSelected"]'
-  );
-
-  const sortDirection = (
-    activeSort.props() as {
-      'data-test-subj': string;
-    }
-  )['data-test-subj'].match(/(?<direction>[^-]+)$/)!.groups!.direction;
-
-  return [columnSorter, sortDirection];
-}
-
 function openColumnSorter(datagrid: ReactWrapper) {
   let popover = datagrid.find(
     'EuiPopover[data-test-subj="dataGridColumnSortingPopover"]'
@@ -198,45 +152,6 @@ function closeColumnSorter(datagrid: ReactWrapper) {
   expect(popover).not.euiPopoverToBeOpen();
 
   return popover;
-}
-
-function sortByColumn(
-  datagrid: ReactWrapper,
-  columnId: string,
-  direction: 'asc' | 'desc' | 'off'
-) {
-  openColumnSorter(datagrid);
-
-  let [columnSorter, currentSortDirection] = getColumnSortDirection(
-    datagrid,
-    columnId
-  );
-
-  // if this column isn't being sorted, enable it
-  if (currentSortDirection === 'off') {
-    act(() => {
-      // @ts-ignore does not require an argument in this usage
-      columnSorter.find('EuiSwitch').props().onChange!();
-    });
-
-    datagrid.update();
-
-    // inspect the column's new sort details
-    [columnSorter, currentSortDirection] = getColumnSortDirection(
-      datagrid,
-      columnId
-    );
-  }
-
-  if (currentSortDirection !== direction) {
-    const sortButton = columnSorter.find(
-      `button[data-test-subj="euiDataGridColumnSorting-sortColumn-${columnId}-${direction}"]`
-    );
-    expect(sortButton.length).toBe(1);
-    sortButton.simulate('click');
-  }
-
-  closeColumnSorter(datagrid);
 }
 
 expect.extend({
@@ -333,10 +248,6 @@ const renderCellValueRowAndColumnCount: RenderCellValue = ({
 
 const renderCellRowAsValue: RenderCellValue = ({ rowIndex }) => rowIndex;
 
-const renderCellValueALowBHigh: RenderCellValue = ({ rowIndex, columnId }) =>
-  // render A as 0, 1, 0, 1, 0 and B as 9->5
-  columnId === 'A' ? rowIndex % 2 : 9 - rowIndex;
-
 describe('EuiDataGrid', () => {
   // Mock requestAnimationFrame to run immediately
   jest
@@ -352,245 +263,6 @@ describe('EuiDataGrid', () => {
       renderCellValue={() => null}
     />
   );
-
-  describe('column sorting', () => {
-    it('calls the onSort callback', () => {
-      const onSort = jest.fn((columns) => {
-        component.setProps({ sorting: { columns, onSort } });
-        component.update();
-      });
-
-      const component = mount(
-        <EuiDataGrid
-          aria-labelledby="#test"
-          columns={[{ id: 'ColumnA' }]}
-          columnVisibility={{
-            visibleColumns: ['ColumnA'],
-            setVisibleColumns: () => {},
-          }}
-          rowCount={1}
-          sorting={{
-            columns: [],
-            onSort,
-          }}
-          renderCellValue={() => 'hello'}
-        />
-      );
-
-      sortByColumn(component, 'ColumnA', 'desc');
-
-      expect(onSort).toHaveBeenCalledTimes(2);
-      expect(onSort).toHaveBeenCalledWith([
-        { id: 'ColumnA', direction: 'asc' },
-      ]);
-      expect(onSort).toHaveBeenCalledWith([
-        { id: 'ColumnA', direction: 'desc' },
-      ]);
-
-      const [, sortDirection] = getColumnSortDirection(component, 'ColumnA');
-      expect(sortDirection).toBe('desc');
-    });
-
-    describe('in-memory sorting', () => {
-      it('sorts on initial render', () => {
-        const renderCellValue: RenderCellValue = ({ rowIndex, columnId }) =>
-          // render A 0->4 and B 9->5
-          columnId === 'A' ? rowIndex : 9 - rowIndex;
-        const component = mount(
-          <EuiDataGrid
-            aria-label="test"
-            columns={[{ id: 'A' }, { id: 'B' }]}
-            columnVisibility={{
-              visibleColumns: ['A', 'B'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={5}
-            renderCellValue={renderCellValue}
-            inMemory={{ level: 'sorting' }}
-            sorting={{
-              columns: [{ id: 'A', direction: 'desc' }],
-              onSort: () => {},
-            }}
-          />
-        );
-
-        expect(extractGridData(component)).toEqual([
-          ['A', 'B'],
-          ['4', '5'],
-          ['3', '6'],
-          ['2', '7'],
-          ['1', '8'],
-          ['0', '9'],
-        ]);
-      });
-
-      it('sorts on multiple columns', () => {
-        const component = mount(
-          <EuiDataGrid
-            aria-label="test"
-            columns={[{ id: 'A' }, { id: 'B' }]}
-            columnVisibility={{
-              visibleColumns: ['A', 'B'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={5}
-            renderCellValue={renderCellValueALowBHigh}
-            inMemory={{ level: 'sorting' }}
-            sorting={{
-              columns: [
-                { id: 'A', direction: 'desc' },
-                { id: 'B', direction: 'asc' },
-              ],
-              onSort: () => {},
-            }}
-          />
-        );
-
-        expect(extractGridData(component)).toEqual([
-          ['A', 'B'],
-          ['1', '6'],
-          ['1', '8'],
-          ['0', '5'],
-          ['0', '7'],
-          ['0', '9'],
-        ]);
-      });
-
-      it('sorts in response to user interaction', () => {
-        const onSort = jest.fn((columns) => {
-          component.setProps({ sorting: { columns, onSort } });
-          component.update();
-        });
-
-        const component = mount(
-          <EuiDataGrid
-            aria-labelledby="#test"
-            columns={[{ id: 'A' }, { id: 'B' }]}
-            columnVisibility={{
-              visibleColumns: ['A', 'B'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={5}
-            renderCellValue={renderCellValueALowBHigh}
-            inMemory={{ level: 'sorting' }}
-            sorting={{
-              columns: [],
-              onSort,
-            }}
-          />
-        );
-
-        expect(extractGridData(component)).toEqual([
-          ['A', 'B'],
-          ['0', '9'],
-          ['1', '8'],
-          ['0', '7'],
-          ['1', '6'],
-          ['0', '5'],
-        ]);
-
-        sortByColumn(component, 'A', 'desc');
-        expect(extractGridData(component)).toEqual([
-          ['A', 'B'],
-          ['1', '8'],
-          ['1', '6'],
-          ['0', '9'],
-          ['0', '7'],
-          ['0', '5'],
-        ]);
-
-        sortByColumn(component, 'B', 'asc');
-        expect(extractGridData(component)).toEqual([
-          ['A', 'B'],
-          ['1', '6'],
-          ['1', '8'],
-          ['0', '5'],
-          ['0', '7'],
-          ['0', '9'],
-        ]);
-      });
-
-      it('sorts with all digit groups in numerical-like', () => {
-        const onSort = jest.fn((columns) => {
-          component.setProps({ sorting: { columns, onSort } });
-          component.update();
-        });
-        const renderCellValue: RenderCellValue = ({ rowIndex }) =>
-          `1.0.${(rowIndex % 3) + rowIndex}`; // computes as 0,2,4,3,5
-
-        const component = mount(
-          <EuiDataGrid
-            aria-label="test"
-            columns={[{ id: 'version' }]}
-            columnVisibility={{
-              visibleColumns: ['version'],
-              setVisibleColumns: () => {},
-            }}
-            rowCount={5}
-            renderCellValue={renderCellValue}
-            inMemory={{ level: 'sorting' }}
-            sorting={{
-              columns: [],
-              onSort,
-            }}
-          />
-        );
-
-        // verify rows are unordered
-        expect(extractGridData(component)).toEqual([
-          ['version'],
-          ['1.0.0'],
-          ['1.0.2'],
-          ['1.0.4'],
-          ['1.0.3'],
-          ['1.0.5'],
-        ]);
-
-        sortByColumn(component, 'version', 'asc');
-
-        expect(extractGridData(component)).toEqual([
-          ['version'],
-          ['1.0.0'],
-          ['1.0.2'],
-          ['1.0.3'],
-          ['1.0.4'],
-          ['1.0.5'],
-        ]);
-      });
-    });
-
-    it('uses schema information to sort', () => {
-      const renderCellValue: RenderCellValue = ({ rowIndex, columnId }) =>
-        // render A 0->4 and B 12->8
-        columnId === 'A' ? rowIndex : 12 - rowIndex;
-      const component = mount(
-        <EuiDataGrid
-          aria-label="test"
-          columns={[{ id: 'A' }, { id: 'B' }]}
-          columnVisibility={{
-            visibleColumns: ['A', 'B'],
-            setVisibleColumns: () => {},
-          }}
-          rowCount={5}
-          renderCellValue={renderCellValue}
-          inMemory={{ level: 'sorting' }}
-          sorting={{
-            columns: [{ id: 'B', direction: 'asc' }],
-            onSort: () => {},
-          }}
-        />
-      );
-
-      expect(extractGridData(component)).toEqual([
-        ['A', 'B'],
-        ['4', '8'],
-        ['3', '9'],
-        ['2', '10'],
-        ['1', '11'],
-        ['0', '12'],
-      ]);
-    });
-  });
 
   describe('updating column definitions', () => {
     it('renders the new set', () => {
